Revert "styles folder and move css files"
This reverts commit 574c65a2c49fb9d24c7ec61b4b358c7396ebaffa.
